progress

+++ src/driver/i2c.h
@@ -32,6 +32,6 @@ uint8_t I2C_Read(bool bFinal);
 int I2C_Write(uint8_t Data);
 
 int I2C_ReadBuffer(void *pBuffer, uint8_t Size);
-int I2C_WriteBuffer(const void *pBuffer, uint16_t size);
+int I2C_WriteBuffer(const void *pBuffer, uint8_t size);
 
 #endif
